Move read_json from slurm_plugin.common to common.utils
