+++ src/slurm_plugin/resume.py
@@ -19,7 +19,8 @@ from logging.config import fileConfig
 
 from botocore.config import Config
 from common.schedulers.slurm_commands import get_nodes_info, set_nodes_down
-from slurm_plugin.common import is_clustermgtd_heartbeat_valid, print_with_count, read_json
+from common.utils import read_json
+from slurm_plugin.common import is_clustermgtd_heartbeat_valid, print_with_count
 from slurm_plugin.instance_manager import InstanceManager
 from slurm_plugin.slurm_resources import CONFIG_FILE_DIR
